+++ RUFAS/biophysical/animal/pen.py
@@ -625,9 +625,39 @@ class Pen:
         specified in `self.manure_streams` and each assigned a `first_processor` directing it how to be routed once
         it reaches the manure module.
         - The function validates that all general stream proportions sum to 1.0 (or 100% of the general portion).
+        - Manure methane potential is assigned according to animal combination. Manure from lactating, dry and close up
+        animals are assigned a value of 0.24 m3 methane per kg of manure volatile solids, and calves and heifers are
+        assigned a value of 0.17, based on the 2024 USDA method for entity scale inventory. If a pen contains heifers,
+        dry and close up animals, methane potential is assigned as a weighted average based on number of dry/close up
+        vs. heifers.
 
         """
         animal_manure_streams: dict[str, ManureStream] = {}
+        if self.animal_combination == AnimalCombination.GROWING_AND_CLOSE_UP:
+            total_animals_in_pen = len(self.animals_in_pen)
+            num_growing = len(
+                [
+                    animal
+                    for animal in self.animals_in_pen.values()
+                    if animal.animal_type in [AnimalType.HEIFER_I, AnimalType.HEIFER_II]
+                ]
+            )
+            num_close_up = len(
+                [
+                    animal
+                    for animal in self.animals_in_pen.values()
+                    if animal.animal_type in [AnimalType.HEIFER_III, AnimalType.DRY_COW]
+                ]
+            )
+            methane_production_potential = (
+                (0.17 * num_growing / total_animals_in_pen + 0.24 * num_close_up / total_animals_in_pen)
+                if total_animals_in_pen > 0
+                else 0.0
+            )
+        else:
+            methane_production_potential = (
+                0.17 if self.animal_combination in [AnimalCombination.CALF, AnimalCombination.GROWING] else 0.24
+            )
 
         pen_animal_excretions = self.total_manure_excretion
         total_pen_manure_data = PenManureData(
@@ -651,6 +681,7 @@ class Pen:
             degradable_volatile_solids=pen_animal_excretions.degradable_volatile_solids,
             total_solids=pen_animal_excretions.total_solids,
             volume=pen_animal_excretions.manure_mass / ManureConstants.SLURRY_MANURE_DENSITY,
+            methane_production_potential=methane_production_potential,
             pen_manure_data=total_pen_manure_data,
         )
 
@@ -732,6 +763,7 @@ class Pen:
             degradable_volatile_solids=manure_stream.degradable_volatile_solids,
             total_solids=manure_stream.total_solids + total_bedding_dry_solids,
             volume=manure_stream.volume + total_bedding_volume,
+            methane_production_potential=manure_stream.methane_production_potential,
             pen_manure_data=manure_stream.pen_manure_data,
         )